Refactor input error types
