+++ Cargo.toml
@@ -12,7 +12,7 @@ name = "cymbalum"
 doctest = false
 
 [dependencies]
-chemfiles = "0.5.4"
+chemfiles = "0.5.5"
 rand = "0.3.14"
 log = "0.3.5"
 toml = "0.1.28"
