+++ src/input/error.rs
@@ -1,0 +1,113 @@
+// Cymbalum, an extensible molecular simulation engine
+// Copyright (C) 2015-2016 G. Fraux — BSD license
+
+use std::io;
+use std::error;
+use std::fmt;
+use std::result;
+
+use toml::Parser;
+use chemfiles;
+
+use units::UnitParsingError;
+
+/// Custom `Result` type for input files
+pub type Result<T> = result::Result<T, Error>;
+
+/// Possible causes of error when reading input files
+#[derive(Debug)]
+pub enum Error {
+    /// Error in the TOML input file
+    TOML(String),
+    /// IO error
+    File(io::Error),
+    /// File content error: missing sections, bad data types
+    Config(String),
+    /// Unit parsing error
+    UnitParsing(UnitParsingError),
+}
+
+impl From<io::Error> for Error {
+    fn from(err: io::Error) -> Error {Error::File(err)}
+}
+
+impl<'a> From<&'a str> for Error {
+    fn from(err: &'a str) -> Error {
+        Error::Config(String::from(err))
+    }
+}
+
+impl From<String> for Error {
+    fn from(err: String) -> Error {
+        Error::Config(err)
+    }
+}
+
+impl From<UnitParsingError> for Error {
+    fn from(err: UnitParsingError) -> Error {Error::UnitParsing(err)}
+}
+
+impl fmt::Display for Error {
+    fn fmt(&self, fmt: &mut fmt::Formatter) -> result::Result<(), fmt::Error> {
+        use std::error::Error as StdError;
+        try!(write!(fmt, "{}", self.description()));
+        Ok(())
+    }
+}
+
+impl error::Error for Error {
+    fn description(&self) -> &str {
+        match *self {
+            Error::TOML(ref err) => err,
+            Error::Config(ref err) => err,
+            Error::File(ref err) => err.description(),
+            Error::UnitParsing(ref err) => err.description(),
+        }
+    }
+
+    fn cause(&self) -> Option<&error::Error> {
+        match *self {
+            Error::TOML(..) => None,
+            Error::Config(..) => None,
+            Error::File(ref err) => Some(err),
+            Error::UnitParsing(ref err) => Some(err),
+        }
+    }
+}
+
+pub fn toml_error_to_string(parser: &Parser) -> String {
+    let nerrors = parser.errors.len();
+    assert!(nerrors > 0);
+
+    let errors = parser.errors.iter().map(|error|{
+        let (line, _) = parser.to_linecol(error.lo);
+        format!("{} at line {}", error.desc, line + 1)
+    }).collect::<Vec<_>>().join("\n    ");
+
+    let plural = if nerrors != 1 {"s"} else {""};
+    return format!("TOML parsing error{}: {}", plural, errors);
+}
+
+/******************************************************************************/
+
+/// Possible error when reading and wrinting to trajectories
+#[derive(Debug)]
+pub struct TrajectoryError(chemfiles::Error);
+
+impl From<chemfiles::Error> for TrajectoryError {
+    fn from(err: chemfiles::Error) -> TrajectoryError {
+        TrajectoryError(err)
+    }
+}
+
+impl fmt::Display for TrajectoryError {
+    fn fmt(&self, fmt: &mut fmt::Formatter) -> result::Result<(), fmt::Error> {
+        self.0.fmt(fmt)
+    }
+}
+
+impl error::Error for TrajectoryError {
+    fn description(&self) -> &str {
+        self.0.description()
+    }
+}
